requirements.txt
fitz
reportlab
streamlit>=1.37.0
openai>=1.40.0
spacy>=3.7.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.0
fpdf2>=2.7.0
